Add files via upload
In this project YOLO Algorithm has been used to train a model which detects the fields like name, age, gender etc in Identity Cards. The implementation of the project is in YOLOv3.ipynb file. The details of the project are stated below.
The dataset consists of 3 different identity cards. These include
Pakistan CNIC
UK Passport
USA Passport
To identify the fields, annotation technique is applied on the dataset. makesense.ai is used for image annotation, which is an open source and free to use image annotator, it annotates the images using AI. All the details and results of the algorithms have been compiled in Data Science - Project Report.pdf